fix: FIX reading data from the pkl file
The script was reading data from only one subject leading to data
redundancy. So the model has an equal accuracy over all subjects
(As we are using Leave-One-Subject-Out cross validation)

Using the pandas read_pickle function instead of our pre-defined
function solves the issue.

Note that this commit does not solve the issue in the read_pkl_file
function in the data_ingestion.py file